Move clients into utilities to avoid circular imports
