+++ src/phoenix/server/app.py
@@ -87,6 +87,7 @@ from phoenix.server.api.schema import schema
 from phoenix.server.grpc_server import GrpcServer
 from phoenix.server.telemetry import initialize_opentelemetry_tracer_provider
 from phoenix.trace.schemas import Span
+from phoenix.utilities.client import PHOENIX_SERVER_VERSION_HEADER
 
 if TYPE_CHECKING:
     from opentelemetry.trace import TracerProvider
@@ -96,8 +97,6 @@ logger = logging.getLogger(__name__)
 router = APIRouter(include_in_schema=False)
 
 templates = Jinja2Templates(directory=SERVER_DIR / "templates")
-
-PHOENIX_SERVER_VERSION_HEADER = "x-phoenix-server-version"
 
 
 class AppConfig(NamedTuple):
@@ -169,9 +168,11 @@ class HeadersMiddleware(BaseHTTPMiddleware):
         request: Request,
         call_next: RequestResponseEndpoint,
     ) -> Response:
+        from phoenix import __version__ as phoenix_version
+
         response = await call_next(request)
         response.headers["x-colab-notebook-cache-control"] = "no-cache"
-        response.headers[PHOENIX_SERVER_VERSION_HEADER] = phoenix.__version__
+        response.headers[PHOENIX_SERVER_VERSION_HEADER] = phoenix_version
         return response
